// .devops/aws/src/imports/k8s.d.ts
export declare class KubeCronJob extends ApiObject {
  /**
   * Returns the apiVersion and kind for "io.k8s.api.batch.v1.CronJob"
   */
  static readonly GVK: GroupVersionKind;
  /**
   * Renders a Kubernetes manifest for "io.k8s.api.batch.v1.CronJob".
   *
   * This can be used to inline resource manifests inside other objects (e.g. as templates).
   *
   * @param props initialization props
   */
  static manifest(props?: KubeCronJobProps): any;
  /**
   * Defines a "io.k8s.api.batch.v1.CronJob" API object
   * @param scope the scope in which to define this object
   * @param id a scope-local name for the object
   * @param props initialization props
   */
  constructor(scope: Construct, id: string, props?: KubeCronJobProps);
  /**
   * Renders the object to Kubernetes JSON.
   */
  toJson(): any;
}
// ...
export interface KubeCronJobProps {
  /**
   * Standard object's metadata. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#metadata
   *
   * @schema io.k8s.api.batch.v1.CronJob#metadata
   */
  readonly metadata?: ObjectMeta;
  /**
   * Specification of the desired behavior of a cron job, including the schedule. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#spec-and-status
   *
   * @schema io.k8s.api.batch.v1.CronJob#spec
   */
  readonly spec?: CronJobSpec;
}
// ...
export interface CronJobSpec {
  /**
   * Specifies how to treat concurrent executions of a Job. Valid values are: - "Allow" (default): allows CronJobs to run concurrently; - "Forbid": forbids concurrent runs, skipping next run if previous run hasn't finished yet; - "Replace": cancels currently running job and replaces it with a new one
   *
   * @schema io.k8s.api.batch.v1.CronJobSpec#concurrencyPolicy
   */
  readonly concurrencyPolicy?: string;
  /**
   * The number of failed finished jobs to retain. Value must be non-negative integer. Defaults to 1.
   *
   * @default 1.
   * @schema io.k8s.api.batch.v1.CronJobSpec#failedJobsHistoryLimit
   */
  readonly failedJobsHistoryLimit?: number;
  /**
   * Specifies the job that will be created when executing a CronJob.
   *
   * @schema io.k8s.api.batch.v1.CronJobSpec#jobTemplate
   */
  readonly jobTemplate: JobTemplateSpec;
  /**
   * The schedule in Cron format, see https://en.wikipedia.org/wiki/Cron.
   *
   * @schema io.k8s.api.batch.v1.CronJobSpec#schedule
   */
  readonly schedule: string;
  /**
   * Optional deadline in seconds for starting the job if it misses scheduled time for any reason.  Missed jobs executions will be counted as failed ones.
   *
   * @schema io.k8s.api.batch.v1.CronJobSpec#startingDeadlineSeconds
   */
  readonly startingDeadlineSeconds?: number;
  /**
   * The number of successful finished jobs to retain. Value must be non-negative integer. Defaults to 3.
   *
   * @default 3.
   * @schema io.k8s.api.batch.v1.CronJobSpec#successfulJobsHistoryLimit
   */
  readonly successfulJobsHistoryLimit?: number;
  /**
   * This flag tells the controller to suspend subsequent executions, it does not apply to already started executions.  Defaults to false.
   *
   * @default false.
   * @schema io.k8s.api.batch.v1.CronJobSpec#suspend
   */
  readonly suspend?: boolean;
}
// ...
export declare class Quantity {
  readonly value: any;
  static fromString(value: string): Quantity;
  static fromNumber(value: number): Quantity;
  private constructor();
}
// ...
export interface ResourceRequirements {
  /**
   * Limits describes the maximum amount of compute resources allowed. More info: https://kubernetes.io/docs/concepts/configuration/manage-resources-containers/
   *
   * @schema io.k8s.api.core.v1.ResourceRequirements#limits
   */
  readonly limits?: {
    [key: string]: Quantity;
  };
  /**
   * Requests describes the minimum amount of compute resources required. If Requests is omitted for a container, it defaults to Limits if that is explicitly specified, otherwise to an implementation-defined value. More info: https://kubernetes.io/docs/concepts/configuration/manage-resources-containers/
   *
   * @schema io.k8s.api.core.v1.ResourceRequirements#requests
   */
  readonly requests?: {
    [key: string]: Quantity;
  };
}